requirements.txt
# Для приложения
fastapi
uvicorn
pydantic
redis
psycopg2-binary
passlib
python-jose
sqlalchemy
python-multipart
dotenv
# Для тестирования
pytest
pytest-mock
httpx
coverage
locust
pytest-cov
bcrypt